backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
chromadb
openai
pydantic
pydantic-settings
python-dotenv
